clients/gmail-mcp/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "permit0-gmail-mcp"
version = "0.1.0"
description = "MCP server exposing Gmail tools, each gated by permit0"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "permit0",
    "mcp>=1.2",
    "google-auth>=2.30",
    "google-auth-oauthlib>=1.2",
    "httpx>=0.27",
]

[project.scripts]
permit0-gmail-mcp = "permit0_gmail_mcp.server:main"

[tool.setuptools.packages.find]
include = ["permit0_gmail_mcp*"]
